+++ src/unix-server.cpp
@@ -84,7 +84,7 @@ static auto parse_arguments(int argc, char** argv) ->
     const auto args = std::span(argv, std::size_t(argc));
 
     for (auto index = optind; index < argc; ++index) {
-        result.emplace_back(args[index]);
+        result.emplace_back(args[static_cast<std::size_t>(index)]);
     }
 
     return result;
